pyproject.toml
[project]
name = "transcribe-yt"
version = "0.1.0"
description = "This is a cool cli tool to transcribe videos"
authors = [
    {name = "pulkit777exe",email = "pulkitpareek88@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "youtube-transcript-api (>=1.0.3,<2.0.0)",
    "google-genai (>=1.8.0,<2.0.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]
